requirements.txt
flask==3.0.3
Flask-Cors==5.0.0
mediapipe==0.10.15
numpy==1.26.4
pillow==11.0.0
gunicorn==23.0.0
